+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [1.0.2] - 2019-07-18
 ### Added
+- Added Numeric => `close?`
+- Added Numeric => `delimit`
 - Added Numeric => `delta`
 - Added String => `ascii_only`
 - Added String => `capitalized?`